NEW FILE move solution search logic into its own file
